Updated the main recipe to just include the `curl` recipe without checks
Updated recipe to always include the curl recipe, and removed the check which was in previously,
since Chef's package install mechanisms should treat an install of an existing package as a no-op.

To clarify, packages are independent of cookbooks - see http://wiki.opscode.com/display/chef/Resources#Resources-Package for details on how package operates, but it basically uses the OS-specific package manager to install the package you specify. Many (most) cookbooks will use package to pull in the stuff they need, but package has no concept of higher-level cookbooks.

Closes #6.