[#26] Increasing the thread stack space (-Xss option) for JVM
This change synchronizes the default value with <https://github.com/elasticsearch/elasticsearch/blob/master/bin/elasticsearch.in.sh>.

The default value is parameterized as a node attribute, allowing override in the node/role configuration.

Closes #26, fixes #25.

Note:
-----

While working with the Sematext folks to get SPM running on our QA
environment, we noticed that SPM would not start with Java 6. After
some investigation, the Sematext folks discovered that SPM would not
start because of the low -Xss setting. This change is to address this.
(see https://github.com/karmi/cookbook-elasticsearch/issues/25)