Remove dependency on Nginx cookbook and use a custom recipe to install it
The [community Nginx cookbook](https://github.com/opscode-cookbooks/nginx)
is effectively FUBAR.

It brings in a dependency on the `bluepill` cookbook, which itself is absolutely FUBAR.
It is not idempotent when installing via source, it is unneccesarily complex,
and keeps breaking all the time.

So a lightweight (43 LOC), custom recipe is used (`recipes/nginx.rb`),
simple configuration files are stored in `templates`.

The `elasticsearch::proxy_nginx` recipe automatically pulls in the `elasticsearch::proxy_nginx`;
if you need more modularity and controll, please submit an issue.

Closes #24, fixes #30.