Added the `logging.yml` template file and corresponding attribute file
Added the ERB template for the `logging.yml` configuration file,
based on the stock Elasticsearch repository one.

Added a simple test case based on attributes set in Vagrantfile.

The configuration handling is dynamic: it walks over all attributes set
in the `node.elasticsearch.logging` namespace, and prints all which are set.

So, to configure eg. the `discovery` level, simply set the corresponding attribute:

    node.elasticsearch.logging['discovery'] = 'TRACE'

As always, set the node attribute directly with `knife` or JSON data for Chef Solo,
with roles, as a part of `settings` databag, etc.

Related: #39