[#77] Added `build-essential` into Vagrantfile and recipes/ebs
NOTE: Even though the "build-essential" cookbook is included in the node `run_list`,
`build-essential` is not installed and `make` is not available before the Chef run,
making the installation of the Fog gem fail.

That's why we need to include the build-essential/make/gcc/etc packages also in the
`value_for_platform_family` in recipes/ebs.