Fixed incorrect "ark" parameters for the installation directory
If you set a custom installation directory using the :dir attribute, the cookbook still installs ElasticSearch in /usr/local/. This is because the call to ark that downloads and installs ElasticSearch doesn't include the necessary parameters to actually place the unpacked files in a custom location.

This patch updates the default cookbook to properly configure ark when using a custom directory.

Closes #80