Update to 0.90.4 RPM
Unsure how you were generating the Checksums for remote_file here, none of mine matched what was in the repository, additionally the checksum above is not the same length as a SHA256 checksum (which is used by remote file).  

I am switching to the SHA256 checksum here, generated by the command `shasum -b -a 256 elasticsearch-0.90.4.noarch.rpm`.