Reduce 'magic' attributes
Remove many of the magic, lazy, or otherwise surprisingly difficult to understand attributes. This shifts the burden to caller to supply more resource parameters when doing something unusual, but makes the default use case easier to understand, primarily by putting default values directly in the resource declared as `default:` instead of making the defaults nil and calculating them in the provider.

Fixes #348.