Dynamically choose user to run plugin commands
Instead of always using the 'elasticsearch' user or the user supplied by the `elasticsearch_user` resource, choose 'root' if we're installing via package, otherwise fall back to the previous logic. This also removes the chown command, since we expect choosing the correct user should get rid of the need for the chown.

Fixes #421.