Allow version upgrade when using "package" install
If the install type is set to package, the install doesn't actually
upgrade to the version we want.  This change updates the method that
runs the package install to include the version when requesting the
package be installed - allowing the upgrade to occur.

This change also adds --force-confold to the package options when using
debian so the configuration changes don't get applied.